+++ website/docs/style/illustration/illustrations-list.js
@@ -33,7 +33,7 @@ const illustartionsList = {
       group: 'States',
     },
     {
-      name: 'MechanismInside',
+      name: 'Configuration',
       group: 'States',
     },
     {