+++ website/docs/.vitepress/theme/DevportalLogo.vue
@@ -1,9 +1,9 @@
 <template>
-  <a href="https://developer.semrush.com/" class="devportal-logo" target="_blank">
+  <a href="https://developer.semrush.com/" class="devportal-logo" target="_blank" aria-label="Developer home page">
     <img src="/devportal-logo.svg" width="30" height="30" class="devportal-logo" alt="Semrush Developer portal" />
     Developer
   </a>
-  <a .href="homeUrl" class="title intergalactic-logo" aria-label="Intergalactic Design System main page">
+  <a .href="homeUrl" class="title intergalactic-logo" aria-label="Intergalactic Design System home page">
     Intergalactic
   </a>
 </template>